+++ tests/framework/db/mssql/connection/ConnectionTest.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yiiunit\framework\db\mssql\command;
+
+use yiiunit\support\MssqlConnection;
+
+/**
+ * @group db
+ * @group mssql
+ * @group connection
+ */
+final class ConnectionTest extends \yiiunit\framework\db\connection\AbstractConnection
+{
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->db = MssqlConnection::getConnection();
+    }
+}